Leave it to the caller to cleanup the connection
The behaviour of the libcurl multi interface is not well documented.

One apparent issue is if you do not wait for all easy handle requests to complete and remove them from the multi handle, libcurl will immediately tear down the connection, sending a TCP RST to the server.

Assuming the calling application implements its own connection pool logic (as FreeRADIUS does), it can create multiple handles, and only cleanup when the handle is next used, by which time the transfer will of hopefully completed.