Use argument of proper type in call to curl_easy_escape()
curl_easy_escape() expects a CURL* so use ->easy[i] and not ->multi
which is a CURLM*.  In the public API both CURL and CURLM are unfortunately
typedef'd to void but internally they're backed by different structs;
struct SessionHandle (urldata.h) and struct Curl_Multi (multihandle.h),
respectively.
This may be a problem because curl_escape_escape() internally calls
Curl_convert_to_network() which operates on fields in struct SessionHandle.